requirements.txt
black==20.8b1
coverage==5.3
flake8==3.8.3
Flask==1.1.2
isort==5.5.4
mypy==0.782
mypy-extensions==0.4.3
pycodestyle==2.6.0
pydocstyle==5.1.1
PyJWT==1.7.1
pylint==2.6.0
requests==2.24.0
unittest2==1.1.0
gunicorn==20.0.4
